chore: component based updates + project re-structuring
