+++ client/pages/dashboard/index.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function UserDashboard() {
+  return <div>UserDashboard</div>;
+}
+
+export default UserDashboard;
