+++ client/pages/sign-up/index.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function SignUp() {
+  return <div> SignUp</div>;
+}
+
+export default SignUp;
