+++ client/trash/all-products.jsx
